[LocalOperator] Add numerical implementations of nonlinear versions of jacobian_apply_*()
This patch adds new mixin classes that provide numerical implementations
of the nonlinear jacobian_apply_() methods. Moreover, it also changes
the existing mixins for the linear versions to inherit from the
nonlinear mixins. That way, inheriting from the linear version gives you
both the linear and the nonlinear mixins. We cannot provide separate
mixins as that will create a name resolution ambiguity due to having two
methods with the same name in two different base classes.