[Documentation] Update local operator interface documentation
This patch clarifies the interface documentation for the linear versions
of jacobian_apply_*() and adds the nonlinear versions of those
methods. In particular, it clarifies that the linear versions only work
due to the fact that the jacobian is constant and describe how to
evaluate the jacobian for the new, nonlinear versions.