Merge branch 'feature/warnings-and-errors-for-sequential-ug' into 'master'
Throw warnings and errors if sequential UG is used in parallel context

A warning is issued at configure time, if such a combination is in use.
An error is thrown through a static assertion if the combination results
in a misleading compile error.

This is mostly a problem when using a distribution-provided UG package
(which tend to lack MPI support). We really should get this cleaned up,
but with this MR users will at least get an error that points them in
the right direction instead of a completely incomprehensible C++
compilation failure.

See merge request !131
(cherry picked from commit 7964ab342599629d3f19871417ec09a80ba86796)

Signed-off-by: Steffen Müthing <muething@dune-project.org>