[cleanup][backend] Sanitize include guards.
This fixes a problem where you would no longer get warnings about deprecated
headers once you fixed the first of those warnings.

Most of the backward compatibility headers contained partial include
guards (i.e. the #define was missing).  Some used the same guard name as the
new headers.  In this patch I opted to upgrade them to full include guards.
This has the disadvantage that in each compile you will only be warned about
the first location that includes the backward compatibility header -- you need
to fix that and recompile to see the next location.

An alternative would have been to remove the include guards entirely in the
backward compatibility headers.  This would have the disadvantage that you get
spammed with warnings, which is a bad thing if you can't do anything about
them at the moment, because it makes other warnings (and errors) harder to
find.