[cleanup] Add include guards in the fem testing snippets.
They are not strictly necessary here, since these seem to be implementation
details of one of the test programs, but they should not hurt either.

N.B. These snippets seem strange.  Their actual code is enclosed in a

  #ifdef USE_<foo>_FEM_FACTORY

The actual define is set from CMake.  I guess the intention is that a type of
finite element map can be selected from dune_add_test() in CMakeLists.txt, and
you don't pay the compilation time for the other finite element map
implementations.

However, all these files have their #include-directives outside of the #ifdef
USE_<foo>_FEM_FACTORY, so all finite element map implementations must be
parsed in any case.  The only thing that is not parsed is the small factory
class.  So it is doubtful that this is effective.