[DGNavierStokes] clean up DG interior penalty usage
The Navier-Stokes-DG parameter class gets rid of the deprecated interior
penalty implementation. An object of the new interior penalty
implementation is now constructed during creation of an instance of the
Navier-Stokes parameter class.
The DG-method is now read from a parameter tree. This implies some
changes in the structure of the ini-file.
Parsing the DG-method from a string is deprecated.
This also implies some changes in the main program code, see the
pdelab/dune-pdelab-howto for an updated version of the main program
code.

The local operators have been modified in order to use the new interior
penalty implementation. The computation of is done outside of the
quadrature loop. The scaling with the viscosity is dropped.

Using now the old interior penalty implementation ***is strongly
discouraged*** as it certainly breaks the code.