[LocalBasisCache] Switch back to mutable cache by default
In order to find a number of bugs in the TBB grid operator (which splits
the local operator), I made the LocalBasisCache non-mutable some time
back. Unfortunately, this creates a lot of major problems with all of
the stock local operators after René switched them to use a cache for
basis evaluation.

This patch thus introduces an optional template paramter on the
cache. If this parameter is set to false (the default), the cache
becomes mutable again and can be used without pre-filling in the context
of a const LocalOperator member function. On the other hand, this is
unsafe in multi-threaded context.

If the parameter is explicitly set to true, the cache cannot be
manipulated when accessed through a const reference.