[WeightedSumLOP] different template parameters in nonlinear jacobian_apply*
... methods for linearization point and input for operator evaluation.

If the types are equal (the standard case in the PDELab machinery) the
call for each local operator goes through the LocalAssemblerCallSwitch.

Otherwise, one example is the usage in the matrix-free preconditioners,
it is assumed that each local operator supports the operator application
for different types of linearization point and input.