Rename namespace 'EIGEN' to 'Eigen'
'Eigen' is not an acronym, therefore it should not be capitalized

Unfortunately, the Eigen library is itself also in a namespace called 'Eigen',
which, when used from within the Dune::PDELab namespace, clashes with Dune::PDELab::Eigen.
Therefore this patch adds quite a number of explicit qualifications.

The old namespace is still there, for backward compatibility.