[backend] Fix container views if DOFIndex and ContainerIndex have same type
The container views allow indexing by both DOFIndex and ContainerIndex;
when using a DOFIndex, it will automatically be mapped before accessing
the data. This arrangement creates a problem when using the views with
function spaces based on dune-functions bases - in this case, the
ordering system is completely bypassed and thus the DOFIndex and the
ContainerIndex are identical. This patch fixes the problem by disabling
the DOFIndex-based functions using SFINAE if both types are identical.