modify 'jacobian_boundary()' to allow usage of CCFV operator for rediscretization in AMG-DG
The old behaviour is the default. But if the operator is used to
discretize the coarse grid problem in AMG-DG, multiply the distance
between the center of the boundary face and the center of the inside
cell by the factor of two. This treats the boundary conditions in the DG
limit correctly.