[istl] Move istl -> ISTL rename compatibility stuff to separate header
We'll have to make sure to always have the contents of the ISTL
namespace injected into the old compatibility space istl, even if users
only include one of the headers in the `istl/` directory. So this patch
moves the code for setting this up into a separate header file and
guards it to avoid repeating the work.

This file will be included by all implementation headers of the ISTL
backend, and to make sure that users in the future only include the
single `istl.hh` header, we also add a check for that. This check can be
temporarily deactivated by defining
`_DUNE_PDELAB_SUPPRESS_ISTL_HH_WARNING` before including a detail
header.

Kudos to Markus for reporting the problem!