[GFS] overwrite all entitysets with the first one
Currently, the grid function spaces are built bottom up, e.g. first the
leaf spaces are constructed and passed to the other nodes and the root
of the gfs tree. By this process, nodes in the gfs tree might have
different entity sets which are then not setup correctly in the gfs
construction process.
This patch overwrites the entity sets of all child spaces of the root
with the entity set of the "leftmost" child prior to creating the
ordering tree. This fixes issues where an entity set was not setup
correctly when the ordering was created, as now all leafs share the same
set. However, with this patch, the entity set object can not be provided
by the user anymore, as it might be overwritten internally.