[CMake] fix Eigen3 test for newer versions (>= 3.0) of CMake
Newer CMake versions don't ship a test for Eigen3 anymore. But the
Eigen3 source code has its own FindEigen3.cmake that can be
used. Therefore extend the CMAKE_MODULE_PATH to the include directory
for the headers.

This works so far on my machine but it hasn't been tested on clusters.