Tell solver backends to reuse matrix when not reassembling.
At least some AMG backends allow to set set a reuse flag which
results in not contruction the AMG including hierarhy from scratch.
We set this flag to true whenever there was no need to reassemble and
otherwise to false.

Note that this implies that setting the parameter with the backend's
constructor has no effect anymore. But that might actually turn out to
be good as previously setting this to true in a nonlinear PDE
meant using a wrong preconditioner throught all steps of the newton
but the first one.