[MatrixFree,Backend] simplify existing linear solvers
The type of on-the-fly operator (linear or nonlinear) does not need to
be specified from the outside anymore. It now takes the grid operator
and the linearized operator that treats things correctly is created
inside the backend classes.

This simplifies the usage of matrix-free nonlinear solvers for both
linear and linearized problems.