[DuneFunctionGFS] initialize EntitySet
When passing a dune-functions basis with a GridView to the
DuneFunctionGFS, the GridView is wrapped in an EntitySet, but the latter
is never updated. This required the user to create and setup the
EntitySet externally and pass it to the dune-functions basis.
This patch creates the EntitySet inside of the DuneFunctionsGFS and
initializes it with all codims, circumventing the above problem. Note
however, that the user cannot use a dune-functions basis with an
EntitySet anymore.