[Bugfix,MatrixFreeNewton] correct computation of current nonlinear reduction
Although this is slightly out of place in this branch / merge request,
it fixes the output and computation of the nonlinear reduction between
the previous and current nonlinear iteration. Before that the output
pointed out that this was equal to the total nonlinear reduction.

Also include some slightly code polishing. In particular:
- use 'not' instead of '!'
- initialize pointers with 'nullptr' instead of '0' or static cast