[bugfix,interpolate] fix #83 and allow interpolation from generic lambdas
we used Dune::Functions::IsCallable, which is part of the
SignatureTraits to check if we are dealing with a callable. This
construct only works if we have a unique operator() (like the while
signature traits can only work in this case). There is an alternative
check as part of the dune-functions concepts
Dune::Functions::Concept::isCallable<F,Domain>. It requires that we know the
parameters Domain and then it tests if F(Domain) is callable. As we
can deduce Domain from the GV passed to makeLocalFunctionTree, we can
now interpolate from a generic lambda.