[fem] Add utility function for calculating the block size of a finite element map
This patch adds a function that calculates the instance-independent
block size of a finite element map.

The main reason for this function is that it can be called in
`constexpr` context with the recent changes to dune-geometry that make
`GeometryType` `constexpr`. If the member function
`FEM::size(GeometryType)` of the finite element map is `static` and
`constexpr`, we can calculate the required block size of the map at
compile time and use it to set the correct block size for ISTL
vectors.