[localfunctionspace] Fix performance bug for composite spaces in fast mode
The FillIndicesVisitor incorrectly processed all child indices in its
afterChild() method, causing a major slowdown for composite spaces in fast mode.
As the indices cannot be used as input to the whole index mapping machinery in
fast mode, this overhead is completely unnecessary, so we can just skip that
step entirely in that mode.