[interpolate] Allow interpolation of vector-valued functions into arbitrary function space trees
This patch relaxes the restriction that vector-valued functions can only be interpolated into a
space of depth 1. Instead, we just make sure that the number of leaves in the function space tree is
identical to the dimension of the range of the vector-valued function.

With this changes, it becomes a lot easier to do interpolation into more complicated spaces like
e.g. Taylor-Hood.