Bugfix: get inverse mass matrix for the _father_ element
Previously, it was the matrix for the current element that was
used.  However, it is multiplied by the coarse element basis,
and it therefore needs to be the matrix of the coarse(=father)
element.

This only makes a difference if current and father element do
not have the same reference element type.  This can happen eg
when doing red-green refinement starting with a grid that
contains cubes.