[FE] add missing size and partial methods
The dune-localfunctions virtual interface for the finite element and
for the local basis contains some functions that were not yet
implemented for the finite elements provided by PDELab. The local
basis virtual interface contains the partial function and the finite
element virtual interface contains the size function. We add these
functions to the finite elements in PDELab.