Do not allow OpenMP to get more than one core in CI
Suitesparse, as installed with Debian, is thread-parallel using OpenMP.
OpenMP silently assumes, it can spawn as many threads as there are cores.
In a worst case scenario, this leads to a number of threads quadratic in
the core count, if you also do parallel test execution with the maximum
number of cores. We solve the issue by disallowing OpenMP to allocate more
than one thread.