Make reading weight of weighted accumulation views const.
The weighted accumulation views of `LocalVector` and `LocalMatrix` used to set
an internal `_modified` flag when *reading* the weight, making the `weight()`
getter method non-const.

This change skips setting the modified flag.  If the underlying vector is
indeed modified through the view, other non-const methods will be used to do
so, and will mark the view as modified.