[bugfix][fastdg] Fix pattern assembly in FastDG
This fixes a bug that originate from a partial exadune port.
In Exadune, the pattern base classes have a fast version, which
was not ported to PDELab. However, in the absence of these, the
PatternEngine of the FastDGGridOperator needs to be more clever
and only operate on one entry of each block.