Add new header common/checks.hh with compile-time switches for additional checks
This new header contains macros for enabling and disabling runtime checks with a
finer granularity than the sledgehammer that is `NDEBUG`. The checks default to
`NDEBUG` if they have not been defined before.

For now, there is only a single macro called `DUNE_PDELAB_ENABLE_CHECK_ASSEMBLY`
for sanity checks in the new assembly infrastructure.