Add new concept ReturnsMatrixPattern
The new assembler machinery returns a `shared_ptr` to a newly constructed
pattern from the new method `matrixPattern()` instead of requiring the user to
create a pattern object, as those cannot be copied or moved due to internal
reference member variables. The new concept makes it possible to detect if a
grid operator supports the new mechanism.