Add overload to assembler that accepts temporary engines
The current assembler wants an lvalue reference to the engine, which is
understandable, but makes it impossible to assemble throwaway engines that are
created as temporary objects. This patch adds a new overload for `assemble()`
that is only active for rvalue references and forwards the object to the normal
implementation.