Factor out LFS flavor handling
This patch moves the rather delicate handling of the LFS flavors into the common
base class `FunctionSpaceProvider`. In doing so, it also simplifies both usage
and implementation (less SFINAE, no need to manually extract the flavor in
downstream code anymore).

The patch also starts the required modifications for FastDG support by turning
on `fast` mode for both the `LocalFunctionSpace` and the `LFSIndexCache` when
enabled in the `Context`.