Add the {fmt} library
For our logging framework, we want to use format strings. The {fmt} library is a
very nice implementation of this functionality, so we will base our logging
framework on it.

While this library is packaged for most distributions, we cannot always rely on
it being available, so we optionally vendor it into PDELab if we cannot find an
installed version of the library. For this purpose, there is now a submodule
that points to a repository mirror on our GitLab server. If the buildsystem does
not detect an installed {fmt} library, we copy the relevant headers to a
directory inside `dune/pdelab/vendor` and update the include paths
appropriately. We don't just install to `fmt/` as that might cause weird clashes
if someone later decides to install the standalone library. The actual source
files are simple compiled into `libdunepdelab`.