Add PDELab-level include for {fmt} with extra functionality
The new header dune/pdelab/logging/fmt.hh will be the canonical way of including
{fmt} in PDELab. In addition to pulling the required upstream headers, it also
adds the user-defined literal `""_fmt`, which turns a regular string literal
into one that can be verified at runtime by {fmt}. The upstream way of doing
this involves wrapping the literal in a call to the macro `fmt()`, but that
pollutes the global namespace with a potentially conflicting macro name.

The literal is defined in `namespace Dune::Literals`, to allow isolated
importing. Moreover, it is injected into `namespace Dune`, so that library code
can use it without additional requirements.

The user-defined literal is a template-based variant, which is not part of the
standard, but supported by all major compilers. Once C++20 comes around, we can
replace it with a fully standard solution. Until then, the build system disables
the compiler warning about using this extension.