Add enum LogLevel and class LogMessage
LogMessage is used to represent the log messages inside the logging system. In
addition to the actual message, it also provides access to additional metadata
like the log level, the time of the message and the used logger.

There are also conversion function to turn a LogLevel into a string and vice
versa.