Make basic logging infrastructure reentrant
When using logging in a multi-threaded program, it would be ideal for most calls
to be reentrant, avoiding expensive and error-prone locking. While it will in
general not be possible to get thread-safe sinks without locking (they normally
have to coordinate parallel access to a shared resource), we can remove some
roadblocks in the front-end part (logger backend and message formatting).

The lazy backend item used in the `PatternFormatSink` needs a buffer for storing
the padded logger for the duration of the formatting operation (in particular,
the storage needs to be valid after exiting the lambda). Previously this was
solved by passing an externally owned buffer to the `backend()` function, but
that is messy because callers have to provide a buffer with sufficient lifetime
and because it becomes difficult to write a re-entrant caller (for e.g. logging
in a multithreaded environment).

So we instead go ahead and create a buffer in the closure, which we then have to
make `mutable`. As the buffer is normally statically allocated, there shouldn't
be any performance issues.

Likewise, `LoggerBackend::handle()` now creates a stack-allocated buffer for
each message instead of using a globally shared buffer instance.