[onestep] Make sure that parallel vectors are consistent after interpolating boundary conditions
The `interpolate()` function on the one step gridoperator does make sure to
preserve unconstrained vector entries across the reinterpolation of constrained
DOFs, but this does not work correctly for constrained overlap values, as those
are constrained on some ranks, but not on others.

This patch fixes the problem by broadcasting overlap data from the rank which
contains it as interior data to the overlap regions.