Backport [!395] Make sure that parallel vectors are consistent after interpolating boundary conditions
Merge branch 'bugfix/make-interpolation-in-onestepgridoperator-consistent' into 'master'

ref:pdelab/dune-pdelab The interpolate() function on the one step gridoperator
does make sure to preserve unconstrained vector entries across the
reinterpolation of constrained DOFs, but this does not work correctly for
constrained overlap values, as those are constrained on some ranks, but not on
others.

This patch fixes the problem by broadcasting overlap data from the rank which
contains it as interior data to the overlap regions.

See merge request [pdelab/dune-pdelab!395]

  [pdelab/dune-pdelab!395]: gitlab.dune-project.org/pdelab/dune-pdelab/merge_requests/395