Fix ambiguous overload for linearizationPoint()
The no-argument version of linearizationPoint() did actually take an argument to
enable a compile-time check for whether the method is enabled (only if the
problem is potentially nonlinear).

this patch changes that method to instead take a bool constant with variable
value inside which lets the compiler disambuate the two overloads.