Really fix GridFunctionSpace fixed size detection when alternating betwenn 0 and non-zero sizes
In order to really fix this bug (see #59, #116, !129), we have to introduce a
separate state for geometry types that have not been visited at all. This patch
adds a constant GT_UNUSED for that purpose. This resolves the ambiguity in the
current code which used a size of 0 to signal that a geometry type has either
not been visited yet or that it actually had size 0 for some entity.