Add array constructors to the power type tree objects
With this change, now it is possible to construct power grid function
spaces and power grid functions with different underlying data.
This is something that was possible with the available constructurs
but not really suitable for a templetized power tree
since the number of arguments had to be hard coded.