Have more CamelCase methods
The Dune style guidelines require camel-case method names.
However, some dune-pdelab methods use lower_case_with_underscores.
This inconsistency is difficult to justify, and therefore
this patch renames several such methods in constraints.hh and
adaptivity.hh to camel-case naming.  To preserve backward-
compatibility, methods with the old names are reintroduced,
and simply forward the renamed-implementations.  These
old-name methods are not deprecated, to make this patch
as uncontroversial as possible.