[vtk] Allow shared_ptr arguments for addSolutionToVTKWriter
This commit introduces an additional signature for addSolutionToVTKWriter
that takes grid function space and data vector as shared pointers.
This is necessary because of the delayed nature of VTK export:
Data sets are registered in a first step and written in a second step.
Currently PDELab implicitly requires both steps to happen in the same scope,
as otherwise the registered function objects are destroyed before written
into a file. The additional signature allows to pass shared pointer objects
that the VTK functions take ownership of. Internally, this required the
BlockVector of the ISTL vector backend to store the grid function space
as a shared pointer.