Allow GridFunctionProbe to be instantiated without function
There is a setter for the function anyway, so it is natural to
be able to construct the probe object without a function.

This is also makes the implementation a bit more consistent with
the rest of PDELab in terms of object ownership: It now stores
a shared pointer and uses stackobject_to_shared_ptr. Before it
was storing a redundant raw pointer.