Matrix-free solver in Newton
Note: The Newton part should be finished but the matrix-free test
fails. Current hypothesis: The test uses numerical Jacobians and numerical
Jacobian applys. This might lead to small changes between the matrix-based and
matrix-free solver which in the end results in a breakdown in the matrix-free
case. If you look at the reduction and defect of the linear solver this seems
to be a possibility since the matrix-based solver is barely able to solve the
problem.