[cmake] Build library as intended with dune_enable_all_packages
When the enable_all_packages feature was introduced, our CMake
minimum requirement did not allow us to add libraries the way
it was intended with the feature. The intermediate solution now
broke for me when testing with installed libraries (it had temporary
build directory paths in its INTERFACE_INCLUDE_DIRECTORIES).
Doing it the intended way fixes the problem.