[bugfix] Adapt nonoverlapping computations to changes in UG
dune-grid has introduced a distinction between the two capabilities
HasEntity and HasEntityIterator several years ago. So far, UG did
not distinguish the two. After recent changes it correctly reflects
UG's capabilities: HasEntity is true for all codims, while HasEntityIterator
is not. In PDELab we used HasEntity to guard iteration, which is not a
correct usage pattern anymore. The commit changes PDELab to use the
HasEntityIterator capability instead.