Use dune-common floating point comparison
Replaced the floating point comparisons with relative epsilon with the
corresponding functions from dune-common to avoid code duplication.

This leads to a slight overhead in the Utility::is_equal function, as
the difference between the numbers is now potentially computed twice.
In the current testing scenario however, this is not noticeable.