Remove DG/FV branch in local basis locks
The branch was originally made to optimize the DG/FV case at compile time since we knew that there was only one lock per local basis. But with entity sets that may not contain an entity (previous commit), we cannot guarantee this anymore. Thus, this commit just removes that branch since the compiler cannot know the number of locks at compile time anymore.